run setup commands
